fix: Update infinite-scroll example to follow skeleton pattern
- Created InfiniteScrollPagination component following the same pattern as skeleton's PaginatedResourceSection
- Moved infinite scroll logic into a reusable component
- Updated collections.$handle.tsx to use the new component with proper typing
- Made the example consistent with skeleton template structure

Note: React 19 type compatibility issues remain (same as skeleton) and will be fixed when React Router v7 adds React 19 support
