+++ examples/infinite-scroll/tsconfig.json
@@ -5,8 +5,10 @@
     "./**/*.ts",
     "./**/*.tsx",
     "../../templates/skeleton/*.d.ts",
-    "../../templates/skeleton/vite.config.ts",
     "../../packages/hydrogen/react-router-patches.d.ts"
+  ],
+  "exclude": [
+    "../../templates/skeleton/vite.config.ts"
   ],
   "compilerOptions": {
     "baseUrl": ".",
